fix avoid deleting requests from archiv with process id 0

+++ src/Zmsdb/Process.php
@@ -381,9 +381,7 @@ class Process extends Base implements Interfaces\ResolveReferences
     {
         if ($process->requests && count($process->requests)) {
             // Beware of resolveReferences=0 to not delete the existing requests
-            if (0 < $process->id) {
-                $this->deleteRequestsForProcessId($process->id);
-            }
+            $this->deleteRequestsForProcessId($process->id);
             $query = new Query\XRequest(Query\Base::INSERT);
             foreach ($process->requests as $request) {
                 if ($request->id >= 0) { // allow deleting requests with a -1 request
@@ -401,9 +399,11 @@ class Process extends Base implements Interfaces\ResolveReferences
 
     protected function deleteRequestsForProcessId($processId)
     {
-        $query =  new Query\XRequest(Query\Base::DELETE);
-        $query->addConditionProcessId($processId);
-        return $this->deleteItem($query);
+        if (0 < $processId) {
+            $query =  new Query\XRequest(Query\Base::DELETE);
+            $query->addConditionProcessId($processId);
+            return $this->deleteItem($query);
+        }
     }
 
     public function readExpiredProcessList(
